+++ tests/Models/User.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Laravel\Fortify\Tests\Models;
+
+class User extends \Illuminate\Foundation\Auth\User
+{
+    protected $table = 'users';
+}
