+++ tests/OrchestraTestCase.php
@@ -4,25 +4,20 @@ namespace Laravel\Fortify\Tests;
 
 use Illuminate\Foundation\Testing\RefreshDatabaseState;
 use Laravel\Fortify\Features;
+use Orchestra\Testbench\Attributes\WithConfig;
 use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase;
 
+#[WithConfig('database.default', 'testing')]
 abstract class OrchestraTestCase extends TestCase
 {
     use WithWorkbench;
 
-    public function setUp(): void
+    protected function withTwoFactorAuthentication($app)
     {
-        if (class_exists(RefreshDatabaseState::class)) {
-            RefreshDatabaseState::$migrated = false;
-        }
-
-        parent::setUp();
-    }
-
-    protected function defineEnvironment($app)
-    {
-        $app['config']->set(['database.default' => 'testing']);
+        $app['config']->set('fortify.features', [
+            Features::twoFactorAuthentication(),
+        ]);
     }
 
     protected function withConfirmedTwoFactorAuthentication($app)
@@ -30,5 +25,16 @@ abstract class OrchestraTestCase extends TestCase
         $app['config']->set('fortify.features', [
             Features::twoFactorAuthentication(['confirm' => true]),
         ]);
+    }
+
+    protected function withoutTwoFactorAuthentication($app)
+    {
+        tap($app['config'], function ($config) {
+            $features = $config->get('fortify.features');
+
+            unset($features[array_search(Features::twoFactorAuthentication(), $features)]);
+
+            $config->set('fortify.features', $features);
+        });
     }
 }
